+++ virvo/virvo/vvsllist.h
@@ -115,7 +115,7 @@ template<class T> class vvSLList
     bool remove();
     bool makeCurrent(int);
     bool find(const T&);
-    int  count() const;
+    size_t  count() const;
     void append(const T&, typename vvSLNode<T>::DeleteType);
     void insertAfter(const T&, typename vvSLNode<T>::DeleteType);
     void insertBefore(const T&, typename vvSLNode<T>::DeleteType);
@@ -430,9 +430,9 @@ template<class T> bool vvSLList<T>::find(const T& x)
 
 //----------------------------------------------------------------------------
 /// Return the number of elements in the list.
-template<class T> int vvSLList<T>::count() const
+template<class T> size_t vvSLList<T>::count() const
 {
-  int numElements = 0;
+  size_t numElements = 0;
   vvSLNode<T>* tmp = head;
 
   while (tmp != NULL)
